Implement basic server with api to database
